DEV: Moved EthereumAddress to enigma-tools-m
